Make controls more precise and movement smoother
Increase FPS to 50, but only allow movement every 4th frame (4 is an
important number: 0.25 is exactly representable in floating point, so we
don’t get rounding errors when adjusting positions on the map).

Only process rockfall every 4th frame too, so that the relative speed of
rocks remains the same.

This means that the keyboard is scanned 50 times per second rather than only
10, so keystrokes aren’t missed. When a keystroke is detected, the frame
cycle is reset, so the movement starts immediately. This affects the
apparent speed of rocks (it slows them down up to a whole frame), but that
doesn’t seem to be noticeable, and it doesn’t change their physics.

Also fix a bug (reported by Ali Anvari) where we could attempt to stop the
slide sound after shutting down PyGame, which gave an error on Windows.