Make a level selector (WIP)
I gave up on pygame-gui at this point. It is quite hard to use as though the
API documentation is extensive, it is not complete, and examples are
lacking.

The use of a custom HTML-like/-subset language plus JSON stylesheets makes
it hard to work out exactly what is possible, allowed and intended.

The last straw was the in the selection list the top pixel of the Acorn font
is chopped off, and I can’t work out why; it seems to be a problem with
UILabels, or with their default theming.