Use the fact that ptext caches fonts to speed up printing
Multiple calls to importlib_resources.as_file are inefficient for a
shrink-wrapped (single file) package, as every time it is called the entire
package is expanded to a temporary directory.

This happens in several places in the program, but at least most of them
only occur once, on start-up. This was the only place where it happened
every frame!