Restore previous organisation of code in __init__ and __main__
It seems that after all the main function must go in __init__, as that is
where the module is loaded from, and __main__ is only used in certain
circumstances.

Follow the setup suggested in
https://docs.python.org/3/library/__main__.html

For now we’ll just have to put up with unused imports not being detected by
pylint in __init__.py for some reason.