Improve rock physics
Alistair Turnbull observed that the old algorithm had a major asymmetry, as
opposed to the bias of rocks rolling left if they can before right: namely,
when multiple rocks could fall into a square, the left-most one always
“won”. Instead, the rock directly above should win.

Fix the physics by scanning the spaces and considering which rocks can fall
into them, rather than the rocks and where they can fall (thanks to Alistair
for the suggestion).

A subtlety is that in order to avoid breaking existing puzzles where moving
up between two columns of rocks causes the left-hand rock to fill the
gap (because the left rock would be scanned first), we must check the
left-hand upper rock before the right-hand upper rock when looking to fill a
gap; that is, when there’s competition for a gap, a rock will roll right,
not left. (Thanks again, Alistair.)